+++ src/processes/routing/screens/payments-screens/services-list-screen.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback } from 'react';
 import { ServicesListContainer } from '@pages/payments/servieces-list';
 import { ServicesListScreenStackProps } from "@processes/routing/types";
-import { PaymentServiceUI } from '@shared/api/payment-categories';
+import { PaymentServiceUI } from '@shared/api/payments';
 
 export const ServicesListScreen: React.FC<ServicesListScreenStackProps> = ({ navigation, route }) => {
     const navigateTo = useCallback((service: PaymentServiceUI) => {
