+++ src/processes/routing/types.ts
@@ -1,6 +1,6 @@
 import { BottomTabScreenProps } from "@react-navigation/bottom-tabs"
 import { NativeStackScreenProps } from "@react-navigation/native-stack"
-import { PaymentServiceUI } from "@shared/api/payment-categories"
+import { PaymentServiceUI } from "@shared/api/payments"
 
 export type AppBottomTabParamList = {
     Main: undefined,
